Created NetSectionsStoreManager
