+++ RoadBuilder/UI/src/mods/ModView/ModView.tsx
@@ -1,16 +1,14 @@
-import { useMemo } from "react";
 import { BottomView } from "../BottomView/BottomView";
 import { SidePanel } from "mods/SidePanel/SidePanel";
 import { DragContextManager } from "mods/Contexts/DragContext";
-import { NetSectionItem } from "domain/NetSectionItem";
 import RB_ClickOnRoad from "images/RB_ClickOnRoad.svg";
 
 import styles from "./ModView.module.scss";
 import { useValue } from "cs2/api";
 import { RoadBuilderToolModeEnum } from "domain/RoadBuilderToolMode";
-import { allNetSections$, roadBuilderToolMode$ } from "mods/bindings";
+import { roadBuilderToolMode$ } from "mods/bindings";
 import ActionPopup from "mods/Components/ActionPopup/ActionPopup";
-import { NetSectionsStoreContext } from "mods/Contexts/NetSectionsStore";
+import { NetSectionsStoreManager } from "mods/Contexts/NetSectionsStore";
 import { RoadPropertiesPanel } from "mods/RoadPropertiesPanel/RoadPropertiesPanel";
 import { LanePropertiesContextManager } from "mods/Contexts/LanePropertiesContext";
 import { useLocalization } from "cs2/l10n";
@@ -18,15 +16,6 @@ import { useLocalization } from "cs2/l10n";
 export const ModView = () => {
   const roadBuilderToolMode = useValue(roadBuilderToolMode$);
   let { translate } = useLocalization();
-  let allNetSections = useValue(allNetSections$);    
-
-  let nStore = useMemo(() => {
-    let nStore = allNetSections.reduce<Record<string, NetSectionItem>>((record: Record<string, NetSectionItem>, cVal: NetSectionItem, cIdx) => {
-      record[cVal.PrefabName] = cVal;
-      return record;
-    }, {});
-    return nStore;
-  }, [allNetSections]);  
 
   let content: JSX.Element | null = null;
   switch (roadBuilderToolMode) {
@@ -57,7 +46,7 @@ export const ModView = () => {
         <>
           <SidePanel />
           <BottomView />
-          <RoadPropertiesPanel />                    
+          <RoadPropertiesPanel />
         </>
       );
       break;
@@ -65,12 +54,12 @@ export const ModView = () => {
       return <></>;
   }
   return (
-    <DragContextManager>    
-      <NetSectionsStoreContext.Provider value={nStore}>
+    <DragContextManager>
+      <NetSectionsStoreManager>
         <LanePropertiesContextManager>
           <div className={styles.view}>{content}</div>
         </LanePropertiesContextManager>
-      </NetSectionsStoreContext.Provider>    
+      </NetSectionsStoreManager>
     </DragContextManager>
   );
 };
